Switched the versions from using the HEAD-jre-snapshot to 30.1.1-jre.
When we merged from the 30.1.1-jre tag, the guava-parent project got the version 30.1.1-jre which was used for release.  But all the sub-projects were still dependent on HEAD-jre-snapshot, which has now moved on beyond 30.1.1.  We want to refer to the version present in this repository, not fetch the latest HEAD snapshot during compilation, so the versions here need to line up.  I'm not sure if it was best to fix the parent to HEAD-jre-snapshot, or the children to 30.1.1-jre.  The latter seems like it will prevent problems with fetching an external snapshot of the wrong version.